+++ lib/ui/pages/board/list_page/widgets/borad_list_card.dart
@@ -1,11 +1,15 @@
 import 'package:ballkkaye_frontend/_core/style/m_color.dart';
 import 'package:ballkkaye_frontend/_core/style/m_icon.dart';
 import 'package:ballkkaye_frontend/_core/style/m_text.dart';
+import 'package:ballkkaye_frontend/data/model/board.dart';
 import 'package:flutter/material.dart';
 
 class BoardListCard extends StatelessWidget {
+  final BoardItem boardItem; // BoardItem 주입
+
   const BoardListCard({
     super.key,
+    required this.boardItem,
   });
 
   @override
@@ -14,7 +18,11 @@ class BoardListCard extends StatelessWidget {
       padding: const EdgeInsets.symmetric(horizontal: 16),
       child: InkWell(
         onTap: () {
-          Navigator.pushNamed(context, "/board/detail");
+          Navigator.pushNamed(
+            context,
+            "/board/detail",
+            arguments: boardItem.boardId,
+          );
         },
         child: Container(
           decoration: BoxDecoration(
@@ -31,14 +39,14 @@ class BoardListCard extends StatelessWidget {
                 Row(
                   mainAxisAlignment: MainAxisAlignment.spaceBetween,
                   children: [
-                    MText.normal5_6('ssar', color: MColor.kLabel.neutral),
-                    MText.normal7_4('15분전', color: MColor.kLabel.normal),
+                    MText.normal5_6(boardItem.nickname, color: MColor.kLabel.neutral),
+                    MText.normal7_4(boardItem.relativeTime, color: MColor.kLabel.normal),
                   ],
                 ),
                 SizedBox(height: 4),
                 // 게시글 제목
                 Text(
-                  '동행후기 써봄 길게길게길게길게길게길게길게길게길게길게길게길게길게',
+                  boardItem.title,
                   style: TextStyle(
                     color: MColor.kLabel.normal,
                     fontSize: 16,
@@ -62,7 +70,7 @@ class BoardListCard extends StatelessWidget {
                             child: MIcon.page.community.likedRed,
                           ),
                         ),
-                        MText.button4_5('20', color: MColor.kLabel.neutral),
+                        MText.button4_5('${boardItem.likeCount}', color: MColor.kLabel.neutral),
                       ],
                     ),
                     SizedBox(width: 4),
@@ -71,7 +79,7 @@ class BoardListCard extends StatelessWidget {
                       children: [
                         MIcon.page.community.comment,
                         SizedBox(width: 4),
-                        MText.button4_5('5', color: MColor.kLabel.neutral),
+                        MText.button4_5('${boardItem.replyCount}', color: MColor.kLabel.neutral),
                       ],
                     ),
                   ],
